Change dev setup commands so that they work
Just tried to set the project up locally on Ubuntu 20.04. `docker-compose exec -f docker-compose.dev.yml ...` does not work and only displays the help text of the `docker-compose exec` command. This is because the `-f <docker-compose-file.yml>`-Flag belongs to the `docker-compose` command. For correct and working commands, please see proposed change.